backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
celery==5.3.4
redis==5.0.1
moviepy==1.0.3
pydub==0.25.1
librosa==0.10.1
pyloudnorm==0.1.1
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.1
WeasyPrint==60.2
markdown==3.5.1
python-dotenv==1.0.0
gunicorn==21.2.0
soundfile==0.12.1
ffmpeg-python==0.2.0
